feat: implement slide configuration and image management for PowerPoint generation
